[Reporting] Refactor report listing UI test (#119016)
* added fast-check baby!

* add new report listing test helpers

* refactored report listing test to be much more concise and focussed in the test file

* fix typo in file name

* remove fast-check dep and fix type issue

* moved non-public specific test code to new common/test folder
