+++ x-pack/plugins/reporting/public/management/__test__/index.ts
@@ -1,0 +1,12 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export { mockJobs } from '../../../common/test';
+
+export { setup } from './report_listing.test.helpers';
+
+export type { TestBed, TestDependencies } from './report_listing.test.helpers';
